内容管理模块，基础部分

+++ xuecheng-plus-content/xuecheng-plus-content-model/src/main/java/com/xuecheng/content/model/po/CourseMarket.java
@@ -30,13 +30,13 @@ public class CourseMarket implements Serializable {
      * 现价
      */
     @TableField(value = "price")
-    private Double price;
+    private Float price;
 
     /**
      * 原价
      */
     @TableField(value = "original_price")
-    private Double originalPrice;
+    private Float originalPrice;
 
     /**
      * 咨询qq
